# temp_openwa/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/labels": {
      "get": {
        "operationId": "LabelController_findAll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labels"
          },
          "400": {
            "description": "Session not ready or not a business account"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all labels (WhatsApp Business only)",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/{labelId}": {
      "get": {
        "operationId": "LabelController_findOne",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label details"
          },
          "404": {
            "description": "Label not found"
          }
        },
        "summary": "Get a specific label by ID",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/chat/{chatId}": {
      "get": {
        "operationId": "LabelController_getChatLabels",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labels for the chat"
          }
        },
        "summary": "Get labels for a specific chat",
        "tags": [
          "labels"
        ]
      },
      "post": {
        "operationId": "LabelController_addLabelToChat",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "labelId": {
                    "type": "string",
                    "description": "Label ID to add"
                  }
                },
                "required": [
                  "labelId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Label added to chat"
          },
          "422": {
            "description": "Labels require a WhatsApp Business account, or the chat type has no labels"
          }
        },
        "summary": "Add a label to a chat",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/chat/{chatId}/{labelId}": {
      "delete": {
        "operationId": "LabelController_removeLabelFromChat",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID to remove",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label removed from chat"
          },
          "422": {
            "description": "Labels require a WhatsApp Business account, or the chat type has no labels"
          }
        },
        "summary": "Remove a label from a chat",
        "tags": [
          "labels"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.12.4",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
...
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}